server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Kanban Board MCP server — FastAPI REST + MCP tools for AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.3",
    "greenlet>=3.3.2",
    "mcp[cli]>=1.27.0",
    "sqlmodel>=0.0.38",
    "uvicorn>=0.43.0",
]

[dependency-groups]
dev = ["pytest-asyncio>=0.23.0", "httpx>=0.28.1", "ruff>=0.9.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
